Native windows compatibility changes: 1) Added handling of dirent (which isn't part of windows) 2) Added missing include of STL files 3) Fixed config bug: name of player wasn't written correctly 4) Removed many warnings caused by implicit castings    - double --> float, e.g. 1.0 --> 1.0f    - size_t --> int,   e.g. karts.size() --> (int)karts.size()
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk@757 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/auto_kart.cpp
@@ -84,7 +84,7 @@ void AutoKart::update (float delta)
 
     const float KART_LENGTH = 1.5f;
 
-    crash_perc += collided ? 3.0f * delta : -0.25 * delta;
+    crash_perc += collided ? 3.0f * delta : -0.25f * delta;
     if(crash_perc < 0.0f) crash_perc = 0.0f;
 
     //Find if any player is ahead of this kart
@@ -445,10 +445,10 @@ void AutoKart::check_crashes(const int STEPS, sgVec3 pos)
         if(crashes.kart == -1) //Don't find a kart to dodge if we got one
             for (size_t j = 0; j < NUM_KARTS; ++j)
             {
-                if(world->getKart(j) == this) continue;
+                if(world->getKart((int)j) == this) continue;
 
                 kart_distance = sgDistanceVec2(step_coord,
-                    world->getKart(j)->getCoord()->xyz);
+                    world->getKart((int)j)->getCoord()->xyz);
 
                 {
                     if(kart_distance < KART_LENGTH + 0.125f * i)
